+++ frontend/src/components/atoms/form/AtomicForm.scss
@@ -1,0 +1,6 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  width: 100%;
+}
