+++ frontend/src/components/atoms/form/index.ts
@@ -1,1 +1,1 @@
-export { default as Form } from './Form';
+export { default as AtomicForm } from './AtomicForm';
